refactor(client): moving the github repos and items to git dir
