+++ packages/amplication-client/src/Application/github/AuthAppWithGithub.tsx
@@ -18,7 +18,7 @@ import { useTracking } from "../../util/analytics";
 import { formatError } from "../../util/error";
 import "./AuthAppWithGithub.scss";
 import CreateRepoDialogContent from "./dialogs/CreateRepoDialogContent/CreateRepoDialogContent";
-import GithubRepos from "./dialogs/GithubRepos";
+import GithubRepos from "./dialogs/GithubRepos/GithubRepos";
 
 type DType = {
   startAuthorizeAppWithGithub: models.AuthorizeAppWithGithubResult;
