+++ packages/amplication-client/src/Application/github/dialogs/CreateRepoDialogContent/CreateRepoDialogContent.tsx
@@ -24,8 +24,6 @@ export default function CreateRepoDialogContent({
       triggerCreation({
         variables: { name: data.name, appId, sourceControlService },
       }).catch((error) => {
-        console.log(error);
-
         setError(error.graphQLErrors[0].message);
       });
       // trackEvent({
