+++ packages/amplication-client/src/Application/github/dialogs/GithubRepos/GithubRepoItem/GithubRepoItem.scss
@@ -1,4 +1,4 @@
-@import "../style/index.scss";
+@import "../../../../../style/index.scss";
 
 .github-repo-item {
   @include flexFullRowWithSpacing;
